antok: Bugfixes for compilation with gcc 4.1.2.
Add a number of virtual destructors, fix one conversion from double to int and
also fix a indentation error. Tested compilation with gcc 4.1.2 and verified
that linking against root libraries compiled with gcc 4.3 works. Thanks to
Daniel Hahne for spotting the problems!

+++ tools/gen_kin.cxx
@@ -287,7 +287,7 @@ void gen_kin(char* infile_name = 0, char* outfile_name = 0, std::string configfi
 	double spl2d = needed / spl1;
 	int spl2;
 	if((spl2d - std::floor(spl2d)) > 0.) {
-		spl2 = std::floor(spl2d + 1.);
+		spl2 = (int)std::floor(spl2d + 1.);
 	} else {
 		spl2 = (int)spl2d;
 	}
